+++ CotEditor/Sources/CEAppDelegate.m
@@ -220,6 +220,7 @@
                                CEDefaultColoringRangeBufferLengthKey: @5000,
                                CEDefaultLargeFileAlertThresholdKey: @(100 * pow(1024, 2)),  // 100 MB
                                CEDefaultAutosavingDelayKey: @5.0,
+                               CEDefaultDocumentConflictOptionKey: @(CEDocumentConflictRevert),
                                };
     [[NSUserDefaults standardUserDefaults] registerDefaults:defaults];